# Improve cursor performance