--- a/src/Parser/Cursor.php
+++ b/src/Parser/Cursor.php
@@ -14,12 +14,14 @@ abstract class Cursor
 
     protected int $peekOffset = 0;
 
-    private string $stringFromIndex;
+    private array $buffer;
+    private int $length;
 
     public function __construct(
-        private string $string,
+        string $string,
     ) {
-        $this->stringFromIndex = $string;
+        $this->buffer = mb_str_split($string);
+        $this->length = count($this->buffer);
     }
 
     public function index(): int
@@ -34,9 +36,21 @@ abstract class Cursor
 
     private function charFromIndex(int $offset = 0): ?string
     {
-        $char = mb_substr($this->stringFromIndex, $offset, 1);
+        $pos = $this->index + $offset;
 
-        if ($char === "\r" && mb_substr($this->stringFromIndex, $offset + 1, 1) === "\n") {
+        if ($pos >= $this->length) {
+            return null;
+        }
+
+        $char = $this->buffer[$pos];
+
+        $nextPos = $pos + 1;
+
+        if (
+            $char === "\r" &&
+            $nextPos < $this->length &&
+            $this->buffer[$nextPos] === "\n"
+        ) {
             return "\n";
         }
 
@@ -58,16 +72,20 @@ abstract class Cursor
         $this->peekOffset = 0;
 
         for ($i = 0; $i < $chars; $i++) {
-            $isClRf = str_starts_with($this->stringFromIndex, "\r\n");
+            $isClRf = $this->substr($this->index, 2) === "\r\n";
 
             $this->incrementIndex($isClRf ? 2 : 1);
         }
     }
 
+    private function substr($start, $length = 1){
+        return implode('', array_slice($this->buffer, $start, $length));
+    }
+
     public function peek(int $chars = 1): ?string
     {
         for ($i = 0; $i < $chars; $i++) {
-            $currentPeek = mb_substr($this->stringFromIndex, $this->peekOffset, 2);
+            $currentPeek = $this->substr($this->index + $this->peekOffset, 2);
 
             $this->peekOffset += $currentPeek === "\r\n" ? 2 : 1;
         }
@@ -90,8 +108,8 @@ abstract class Cursor
     public function slice(int $start, int $length = 1): string
     {
         return $start >= $this->index
-            ? mb_substr($this->stringFromIndex, $start - $this->index, $length)
-            : mb_substr($this->string, $start, $length);
+            ? $this->substr($start - $this->index, $length)
+            : $this->substr($start, $length);
     }
 
     public function peekBlankInline(): void
@@ -109,7 +127,7 @@ abstract class Cursor
 
         $this->peekBlankInline();
 
-        $blank = mb_substr($this->stringFromIndex, $start, $this->peekOffset - $start);
+        $blank = $this->substr($this->index + $start, $this->peekOffset - $start);
 
         $this->skipToPeek();
 
@@ -215,14 +233,10 @@ abstract class Cursor
     protected function setIndex(int $index): void
     {
         $this->index = $index;
-
-        $this->stringFromIndex = mb_substr($this->string, $index);
     }
 
     private function incrementIndex(int $offset = 1): void
     {
         $this->index += $offset;
-
-        $this->stringFromIndex = mb_substr($this->stringFromIndex, $offset);
     }
 }